move all tests to parquet/tests/arrow_reader/statistics.rs, and leave a comment in original file
